refactor: Reorganize file structure for better separation of concerns
Organized files for in-course tabs, deleted unnecessary components, refactored message post req, moved reusable function from hooks to lib and renamed to apiHelper.ts
