+++ client/src/pages/CourseView/CourseHeader.tsx
