+++ client/src/pages/Dashboard.tsx
@@ -4,11 +4,9 @@ import { useQuery } from "@tanstack/react-query";
 import { PlusCircle } from "react-feather";
 import { Link } from "react-router-dom";
 
-import { NAME_PLACEHOLDER } from "../components/dummies";
-
 import EmptyFallback from "../components/EmptyFallback";
 import Loading from "../components/Loading";
-import CourseList from "../components/CourseList";
+import CourseList from "./CourseView/CourseList";
 
 const Dashboard = () => {
   const { getCourses } = useCoursesApi();
@@ -23,7 +21,7 @@ const Dashboard = () => {
       {/* Welcome Section */}
       <div className="bg-gradient-to-r from-primary-50 to-primary-100 rounded-xl p-8 shadow-sm border border-primary-200">
         <div className="max-w-3xl">
-          <h1 className="text-2xl md:text-3xl font-bold text-gray-800">Welcome back, <span className="text-primary-600">{NAME_PLACEHOLDER}</span>!</h1>
+          <h1 className="text-2xl md:text-3xl font-bold text-gray-800">Welcome back, <span className="text-primary-600">USER</span>!</h1>
           <p className="mt-3 text-gray-600 md:text-lg">Pick up where you left off or start something new.</p>
           
           <div className="mt-6 flex flex-wrap gap-4">
