+++ client/src/pages/QuizView/QuizCard.tsx
