+++ client/src/pages/QuizView/QuizHeader.tsx
