+++ client/src/services/chatbot.ts
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import { useApiClient } from "./api";
+import callApi from "../lib/apiHelper";
 
 interface Message {
     previous_messages: {
@@ -13,20 +14,10 @@ export const useChatbot = () => {
   const api = useApiClient();
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
-  
+
   const sendMessage = async (message: Message, courseId: number) => {
-    setLoading(true);
-    setError(null);
-    try {
-      const response = await api.post(`api-courses/chat/${courseId}/`, message);
-      return response.data;
-    } catch (err: any) {
-      setError(err.message || 'Something went wrong');
-      throw err;
-    } finally {
-      setLoading(false);
-    }
-  }
+    return callApi(() => api.post(`api-courses/courses/${courseId}/chat/`, message), setLoading, setError);
+  };
 
   return { sendMessage, loading, error };
 }
